requirements.txt
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.16
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mongoengine==0.20.0
oauthlib==3.1.0
Pillow==7.2.0
pkg-resources==0.0.0
pymongo==3.11.0
PySocks==1.7.1
python-daemon==2.2.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3
